feat: Change type of `Guild.bitrate_limit` to `int`.

+++ discord/guild.py
@@ -823,14 +823,14 @@ class Guild(Hashable):
         )
 
     @property
-    def bitrate_limit(self) -> float:
+    def bitrate_limit(self) -> int:
         """The maximum bitrate for voice channels this guild can have."""
         vip_guild = (
             self._PREMIUM_GUILD_LIMITS[1].bitrate
             if "VIP_REGIONS" in self.features
             else 96e3
         )
-        return max(vip_guild, self._PREMIUM_GUILD_LIMITS[self.premium_tier].bitrate)
+        return int(max(vip_guild, self._PREMIUM_GUILD_LIMITS[self.premium_tier].bitrate))
 
     @property
     def filesize_limit(self) -> int:
